+++ deal/__init__.py
@@ -1,2 +1,4 @@
-from .deal import *
+from .core import DataConfig, DEALConfig, FlareConfig, DEAL
+
+__all__ = ["DataConfig", "DEALConfig", "FlareConfig", "DEAL"]
 from .utils import *
