+++ pyproject.toml
@@ -1,0 +1,33 @@
+[project]
+name = "deal"
+description = "Data Efficient Active Learning"
+dynamic = ["version"]
+readme = "README.md"
+authors = [
+    { name = "Luigi Bonati" , email = "luigi.bonati@iit.it"},
+]
+license = { text = "MIT" }
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Natural Language :: English",
+]
+requires-python = ">=3.8"
+dependencies = [
+    "ase",
+    "pandas",
+    "chemiscope",
+]
+
+[project.urls]
+"Source" = "https://github.com/luigibonati/DEAL/"
+
+[project.scripts]
+deal = "deal.cli:main"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["deal"]
